revise calculate_sircle

+++ chapter1/calcurate_circle/main.c
@@ -2,9 +2,9 @@
 #include <math.h>
 
 int main() {
-    float r,S; // 半径,円の面積
+    double r,S; // 半径,円の面積
     printf("r: ");
-    scanf("%f", &r);
+    scanf("%lf", &r);
     S = r * r * M_PI; 
-    printf("S: %f\n", S); // 少数第六位まで表示
+    printf("S: %lf\n", S); // 少数第六位まで表示
 }
